+++ quill/include/quill/std/ForwardList.h
@@ -21,7 +21,7 @@
   #include <string>
 #endif
 
-namespace quill::detail
+namespace quill
 {
 /***/
 template <typename T, typename Allocator>
@@ -136,7 +136,7 @@ struct Decoder<std::forward_list<T, Allocator>,
       for (size_t i = 0; i < number_of_elements; ++i)
       {
         std::wstring_view v = Decoder<T>::decode(buffer, nullptr);
-        encoded_values.emplace_back(utf8_encode(v));
+        encoded_values.emplace_back(detail::utf8_encode(v));
       }
 
       args_store->push_back(encoded_values);
@@ -144,4 +144,4 @@ struct Decoder<std::forward_list<T, Allocator>,
   }
 };
 #endif
-} // namespace quill::detail
+} // namespace quill